+++ php/tests/Application/Command/ClassInfoTest/MethodImplementation.php
@@ -19,12 +19,12 @@ interface TestInterface
 
 class ChildClass extends ParentClass implements TestInterface
 {
-    public function parentInterfaceMethod()
+    public function parentInterfaceMethod(Foo $foo = null)
     {
 
     }
 
-    public function interfaceMethod()
+    public function interfaceMethod(Foo $foo = null)
     {
 
     }